\name{pblm-package}
\alias{pblm-package}
\docType{package}
\title{
Bivariate Additive Marginal Logistic Regression via Maximum Penalized Likelihood Estimation}
\description{Fitting bivariate additive marginal logistic regression models for  categorical responses.}
\details{
\tabular{ll}{
Package: \tab pblm\cr
Type: \tab Package\cr
Version: \tab 0.1-8\cr
Date: \tab 2017-04-02\cr
License: \tab GPL (>= 2)\cr
}
It is possible to fit partial proportional odds models and to specify several parametrizations for the marginals and the association, including the Dale's model. P-splines can be used to smooth covariates. The association structure can also be smoothed by polynomial surfaces by using penalty terms.
}
\author{
Marco Enea, with contributions by Mikis Stasinopoulos and Robert Rigby

Maintainer: Marco Enea <marco.enea@libero.it>
}

